Moved model stuff in model folder and cleaned up imports
