+++ src/app/components/Login/Login.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Flex, Heading, Text, Image, Well } from '@adobe/react-spectrum';
 
-import { setCredentials } from '../../../model/AuthState';
+import { setCredentials } from '../../../model/authState/AuthState';
 
 import logo from '../../media/logo.png';
-import LoginForm from './LoginForm';
 import useAuthState from '../../../hooks/useAuthState';
+import LoginForm from './LoginForm';
 
 export default function Login() {
 	const [{ error }, dispatch] = useAuthState();
