+++ src/app/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Login from '../components/Login';
+import Login from '../components/Login/Login';
 
 export default function LoginPage() {
 	return <Login />;
