+++ src/lib/vertxEventBus/InvalidSocketStateError.ts
