+++ src/model/appSettings/AppSettingsContext.ts
@@ -1,7 +1,7 @@
 import { createContext, Dispatch } from 'react';
 
+import AppSettings, { AppSettingsAction } from './AppSettings';
 import defaultAppSettings from './defaultAppSettings';
-import AppSettings, { AppSettingsAction } from '../model/AppSettings';
 
 const AppSettingsContext = createContext<
 	[AppSettings, Dispatch<AppSettingsAction>]
