+++ src/model/authState/AuthStateContext.ts
@@ -1,5 +1,6 @@
 import { createContext, Dispatch } from 'react';
-import AuthState, { AuthStateAction } from '../model/AuthState';
+
+import AuthState, { AuthStateAction } from './AuthState';
 import initialAuthState from './initialAuthState';
 
 const AuthStateContext = createContext<[AuthState, Dispatch<AuthStateAction>]>([
