+++ src/model/connection/ConnectionContext.ts
@@ -1,6 +1,6 @@
 import { createContext, Dispatch } from 'react';
 
-import Connection, { ConnectionAction } from '../model/Connection';
+import Connection, { ConnectionAction } from './Connection';
 import initialConnection from './initialConnection';
 
 const ConnectionContext = createContext<
