+++ src/model/dashboard/WidgetList.ts
@@ -2,6 +2,5 @@ import { FunctionComponent } from 'react';
 import WidgetProps from './WidgetProps';
 
 export default interface WidgetList {
-	// TODO: Specify props for Widget nodes
 	[key: string]: FunctionComponent<WidgetProps>;
 }
