+++ src/model/dashboard/WidgetProps.ts
