+++ src/model/vertxEventBus/Channels.ts
