+++ src/model/vertxEventBus/JSONSerializable.ts
