+++ src/model/vertxEventBus/VertxEventBus.ts
