+++ src/widgets/ProcedureWidget/ProcedureWidget.tsx
