+++ src/widgets/hooks/useChannelLatest.ts
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
 
-import { Channel } from '../../model/Channels';
-import JSONSerializable from '../../model/JSONSerializable';
+import { Channel } from '../../model/vertxEventBus/Channels';
+import JSONSerializable from '../../model/vertxEventBus/JSONSerializable';
 
 import useChannel from './useChannel';
